requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
playwright==1.47.0
python-dateutil
httpx>=0.27.0

# AutoGen (ag2) – multi-agent framework
autogen-agentchat==0.7.5
autogen-ext[ollama]==0.7.5
autogen-core==0.7.5